Moving files to be encapsulated inside of a single project
